+++ build.ini
@@ -1,6 +1,6 @@
 [Application]
 name=Zygans Parabox Editor
-version=1.10
+version=1.9.4
 entry_point=sdl2_start:main
 icon=icon.ico
 
@@ -9,7 +9,7 @@ entry_point=glfw_start:main
 icon=icon.ico
 
 [Python]
-version=3.10.11
+version=3.10.4
 
 [Include]
 pypi_wheels=PyOpenGL==3.1.6
@@ -19,11 +19,15 @@ pypi_wheels=PyOpenGL==3.1.6
 ;Build from https://github.com/sudo-jade/pyimgui
 local_wheels=imgui-1.4.1-cp310-cp310-win_amd64.whl
 files=level.py > $INSTDIR\pkgs
-      editor\__init__.py > $INSTDIR\pkgs\editor
-      editor\loadsave.py > $INSTDIR\pkgs\editor
+      editor.py > $INSTDIR\pkgs
       style.py > $INSTDIR\pkgs
       hubtools.py > $INSTDIR\pkgs
       state.py > $INSTDIR\pkgs
       style.py > $INSTDIR\pkgs
-      place > $INSTDIR\pkgs
+      place\block.py > $INSTDIR\pkgs\place
+      place\floor.py > $INSTDIR\pkgs\place
+      place\palette.py > $INSTDIR\pkgs\place
+      place\ref.py > $INSTDIR\pkgs\place
+      place\utils.py > $INSTDIR\pkgs\place
+      place\walls.py > $INSTDIR\pkgs\place
