+++ state.py
@@ -1,4 +1,3 @@
-from typing import Any
 class UIstate:
     focused = False
 class Design:
@@ -33,24 +32,4 @@ class UsefulMod:
             for tag in ["Buttont", "PlayerButtont"]:
                 if tag not in floor_types:
                     floor_types.append(tag)
-"""
-class _FrameDebug(dict):
-    def __init__(self):
-        dict.__init__(self)
-        self._print = []
-    def ensure(self, __name: str, default: Any=None) -> None:
-        if not (__name in self): 
-            self[__name] = default
-    def __getattr__(self, __name: str) -> None:
-        return self[__name]
-    def __setattr__(self, __name: str, __value: Any) -> None:
-        self[__name] = __value
-    def print(self,code):
-        self._print.append(str(code))
-    def printload(self):
-        ret = '\n'.join(self._print)
-        self._print = []
-        return(ret)
-FrameDebug=_FrameDebug()
-"""
 usefulmod = UsefulMod()
